Reorganize socket handler and re-create user on server side
